refactor:css variable rename
