+++ src/App.tsx
@@ -2,9 +2,10 @@ import { useState } from "react";
 import Home from "./pages/Home";
 
 import { createContext } from "react";
-import themes, { Theme } from "./constant/themes.constant";
+import THEMES from "./constant/themes.constant";
 import ThemeToggle from "./components/ThemeToggle";
 import Contact from "./components/Contact";
+import { Theme } from "./types";
 
 const stateTheme = localStorage.getItem(
   "react_todo_theme_Id"
@@ -12,10 +13,10 @@ const stateTheme = localStorage.getItem(
 
 function initFilterThemes(): Theme[] {
   if (stateTheme) {
-    const filter = themes.filter((i) => i.themeId !== stateTheme);
+    const filter = THEMES.filter((i) => i.themeId !== stateTheme);
     return filter;
   } else {
-    return themes;
+    return THEMES;
   }
 }
 
@@ -36,7 +37,7 @@ function App() {
   function toggleTheme(themeId: Theme["themeId"]) {
     localStorage.setItem("react_todo_theme_Id", themeId);
     setTheme(themeId);
-    filterThemes(themes.filter((i) => i.themeId !== themeId));
+    filterThemes(THEMES.filter((i) => i.themeId !== themeId));
   }
 
   return (
