+++ src/constant/css.constant.ts
@@ -1,13 +1,10 @@
 // import { ThemeColor } from "./themes.constant";
 
-export type ComponentsStatus = "error";
-export const globalVariableStatus: Record<ComponentsStatus, string> = {
+import { CSSStatus } from "../types";
+
+export const GLOBAL_VAR_STATUS: Record<CSSStatus, string> = {
   error: "248,113,113",
 };
-
-export interface themeVariable {
-  primary: string;
-}
 
 // export const globalVariableTheme: Record<ThemeColor, themeVariable> = {
 //     green: {
