+++ src/dynamicVariables.ts
@@ -1,12 +1,7 @@
-import {
-  ComponentsStatus,
-  globalVariableStatus,
-} from "./constant/css.constant";
+import { GLOBAL_VAR_STATUS } from "./constant/css.constant";
+import { CSSStatus } from "./types";
 const root = document.getElementById("root");
 
-for (const key in globalVariableStatus) {
-  root?.style.setProperty(
-    `--${key}`,
-    globalVariableStatus[key as ComponentsStatus]
-  );
+for (const key in GLOBAL_VAR_STATUS) {
+  root?.style.setProperty(`--${key}`, GLOBAL_VAR_STATUS[key as CSSStatus]);
 }
